Rename confparser to zdlconf
