+++ CMakeLists.txt
@@ -54,13 +54,12 @@ add_executable(
 	libwad.cpp
 	qzdl.cpp
 	${ZDL_MOC_OUTFILES}
-
-	${PROJECT_SOURCE_DIR}/zdlconf/confparser.cpp
-	${PROJECT_SOURCE_DIR}/zdlconf/inih-r42/ini.c
+	${PROJECT_SOURCE_DIR}/zdlconf/zdlconf.cpp
+	${PROJECT_SOURCE_DIR}/zdlconf/inih-r44/ini.c
 )
 
 target_include_directories(zdl PRIVATE ${PROJECT_SOURCE_DIR})
 target_include_directories(zdl PRIVATE ${PROJECT_SOURCE_DIR}/zdlconf)
-target_include_directories(zdl PRIVATE ${PROJECT_SOURCE_DIR}/zdlconf/inih-r42)
+target_include_directories(zdl PRIVATE ${PROJECT_SOURCE_DIR}/zdlconf/inih-r44)
 target_link_libraries(zdl Qt5::Core Qt5::Widgets)
 target_compile_definitions(zdl PRIVATE FALSE=0 TRUE=1)
