+++ ZDLAdvancedMultiplayerDialog.cpp
@@ -1,23 +1,23 @@
 /*
  * This file is part of qZDL
  * Copyright (C) 2007-2010  Cody Harris
- * 
+ *
  * qZDL is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
- 
+
 #include "ZDLAdvancedMultiplayerDialog.h"
-#include "confparser.h"
+#include "zdlconf.h"
 #include <QDialogButtonBox>
 
 ZDLAdvancedMultiplayerDialog::ZDLAdvancedMultiplayerDialog(ZDLWidget *parent):QDialog(parent){
@@ -25,17 +25,17 @@ ZDLAdvancedMultiplayerDialog::ZDLAdvancedMultiplayerDialog(ZDLWidget *parent):QD
 	QVBoxLayout *box = new QVBoxLayout(this);
 	QGridLayout *form = new QGridLayout();
 	QDialogButtonBox *btnBox = new QDialogButtonBox(QDialogButtonBox::Ok|QDialogButtonBox::Cancel,Qt::Horizontal,this);
-	
+
 	extratic = new QCheckBox("On/Off", this);
-	netmode = new QComboBox(this);	
+	netmode = new QComboBox(this);
 	portNo = new QLineEdit(this);
 	dupmode = new QComboBox(this);
 	enable = new QCheckBox("Yes/No", this);
-		
+
 	netmode->addItem("Not Specified");
 	netmode->addItem("0 (Classic)");
 	netmode->addItem("1 (Client/Server Model)");
-	
+
 	dupmode->addItem("Not Specified");
 	dupmode->addItem("1");
 	dupmode->addItem("2");
@@ -46,10 +46,10 @@ ZDLAdvancedMultiplayerDialog::ZDLAdvancedMultiplayerDialog(ZDLWidget *parent):QD
 	dupmode->addItem("7");
 	dupmode->addItem("8");
 	dupmode->addItem("9");
-	
+
 	connect(btnBox, &QDialogButtonBox::accepted, this, &ZDLAdvancedMultiplayerDialog::save);
 	connect(btnBox, &QDialogButtonBox::rejected, this, &ZDLAdvancedMultiplayerDialog::close);
-	
+
 	form->addWidget(new QLabel("Extratic:",this),0,0);
 	form->addWidget(extratic,0,1);
 	form->addWidget(new QLabel("Net Mode:",this),1,0);
@@ -60,22 +60,22 @@ ZDLAdvancedMultiplayerDialog::ZDLAdvancedMultiplayerDialog(ZDLWidget *parent):QD
 	form->addWidget(dupmode,3,1);
 	form->addWidget(new QLabel("Enable:",this),4,0);
 	form->addWidget(enable,4,1);
-	
+
 	QLabel *top = new QLabel("ZDL",this);
 	QFont font;
 	font.setPointSize(32);
 	top->setFont(font);
 	top->setAlignment(Qt::AlignHCenter);
-	
+
 	QLabel *adv = new QLabel("Advanced Multiplayer Settings", this);
 	adv->setAlignment(Qt::AlignHCenter);
-	
+
 	QLabel *messageOne = new QLabel("If no port is given, the default will be used.", this);
 	QLabel *messageTwo = new QLabel("Settings will only be applied to a multiplayer game and when enabled.", this);
 	messageOne->setWordWrap(true);
 	messageTwo->setWordWrap(true);
-	
-	
+
+
 	box->addWidget(top);
 	box->addWidget(adv);
 	box->addLayout(form);
@@ -95,7 +95,7 @@ void ZDLAdvancedMultiplayerDialog::readConfig(){
 	if(zconf->contains("zdl.net/advenabled")){
 		tEnabled = zconf->value("zdl.net/advenabled").toString();
 	}
-	
+
 	QString tExtratic = "disabled";
 	if(zconf->contains("zdl.net/extratic")){
 		tExtratic = zconf->value("zdl.net/extratic").toString();
@@ -115,21 +115,21 @@ void ZDLAdvancedMultiplayerDialog::readConfig(){
 	if(zconf->contains("zdl.net/dup")){
 		tDup = zconf->value("zdl.net/dup").toString();
 	}
-		
+
 	if(tEnabled == "enabled"){
 		enable->setCheckState(Qt::Checked);
 	}else{
 		enable->setCheckState(Qt::Unchecked);
 	}
-	
+
 	if(tExtratic == "enabled"){
 		extratic->setCheckState(Qt::Checked);
 	}else{
 		extratic->setCheckState(Qt::Unchecked);
 	}
-	
+
 	portNo->setText(tPort);
-	
+
 	bool ok;
 	int nDup = tDup.toInt(&ok, 10);
 	int nNetMode = tNetMode.toInt(&ok, 10);
@@ -138,7 +138,7 @@ void ZDLAdvancedMultiplayerDialog::readConfig(){
 	}else if(nDup <= 0){
 		nDup = 0;
 	}
-	
+
 	if(nNetMode > 2){
 		nNetMode = 2;
 	}else if(nNetMode < 0){
@@ -160,7 +160,7 @@ void ZDLAdvancedMultiplayerDialog::save(){
 	}else{
 		zconf->setValue("zdl.net/extratic", "disabled");
 	}
-	
+
 	if(portNo->text().length() > 0){
 		zconf->setValue("zdl.net/port", portNo->text());
 	}else{
@@ -168,8 +168,8 @@ void ZDLAdvancedMultiplayerDialog::save(){
 			zconf->remove("zdl.net/port");
 		}
 	}
-	
-	
+
+
 	if(dupmode->currentIndex() > 0){
 		zconf->setValue("zdl.net/dup", QString::number(dupmode->currentIndex()));
 	}else{
@@ -177,7 +177,7 @@ void ZDLAdvancedMultiplayerDialog::save(){
 			zconf->remove("zdl.net/dup");
 		}
 	}
-	
+
 	if(netmode->currentIndex() > 0){
 		zconf->setValue("zdl.net/netmode", QString::number(netmode->currentIndex()));
 	}else{
@@ -185,8 +185,8 @@ void ZDLAdvancedMultiplayerDialog::save(){
 			zconf->remove("zdl.net/netmode");
 		}
 	}
-	
-	
+
+
 	done(1);
 }
