+++ ZDLIWadList.cpp
@@ -1,24 +1,24 @@
 /*
  * This file is part of qZDL
  * Copyright (C) 2007-2010  Cody Harris
- * 
+ *
  * qZDL is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
- 
+
 #include "ZDLIWadList.h"
 #include "ZDLNameListable.h"
-#include "confparser.h"
+#include "zdlconf.h"
 #include "ZDLNameInput.h"
 
 
@@ -92,7 +92,7 @@ void ZDLIWadList::addButton(){
          << "pk3 Files (*.pk3)"
          << "zip Files (*.zip)"
          << "All files (*)";
-	
+
 	ZDLNameInput diag(this);
 	diag.setWindowTitle("Add IWAD");
 	diag.setFilter(filters);
@@ -123,7 +123,7 @@ void ZDLIWadList::editButton(QListWidgetItem * item){
 			QString name = diag.getName();
 			zitem->setDisplayName(name);
 			zitem->setFile(fileName);
-		}	
+		}
 	}
 }
