+++ ZDLSettingsPane.cpp
@@ -1,17 +1,17 @@
 /*
  * This file is part of qZDL
  * Copyright (C) 2007-2010  Cody Harris
- * 
+ *
  * qZDL is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -21,7 +21,7 @@
 #include <QComboBox>
 #include "ZDLMapFile.h"
 
-#include "confparser.h"
+#include "zdlconf.h"
 #include "ZDLSettingsPane.h"
 
 ZDLSettingsPane::ZDLSettingsPane(QWidget *parent):ZDLWidget(parent){
@@ -267,7 +267,7 @@ void ZDLSettingsPane::newConfig(){
 			if(set == 0){
 				sourceList->setCurrentIndex(0);
 			}
-		}	
+		}
 	}
 
 	IWADList->clear();
@@ -293,7 +293,7 @@ void ZDLSettingsPane::newConfig(){
 					set = 1;
 					IWADList->setCurrentRow(i);
 					break;
-				}	
+				}
 			}
 		}
 		if(!set){
