+++ ZDLSourcePortList.cpp
@@ -1,24 +1,24 @@
 /*
  * This file is part of qZDL
  * Copyright (C) 2007-2010  Cody Harris
- * 
+ *
  * qZDL is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
- 
+
 #include "ZDLSourcePortList.h"
 #include "ZDLNameListable.h"
-#include "confparser.h"
+#include "zdlconf.h"
 #include "ZDLNameInput.h"
 
 #include <cstdio>
@@ -50,14 +50,14 @@ void ZDLSourcePortList::rebuild(){
 	zconf->beginGroup("zdl.ports");
 	zconf->remove("");
 	zconf->endGroup();
-	
+
 	for(int i = 0; i < count(); i++){
 		QListWidgetItem *itm = pList->item(i);
 		ZDLNameListable* fitm = (ZDLNameListable*)itm;
 		zconf->setValue(QString("zdl.ports/p%1n").arg(i), fitm->getName());
 		zconf->setValue(QString("zdl.ports/p%1f").arg(i), fitm->getFile());
 	}
-	
+
 }
 
 void ZDLSourcePortList::newDrop(QStringList fileList){
@@ -92,8 +92,8 @@ void ZDLSourcePortList::addButton(){
 	filters << "Application (*.app)";
 #endif
 	filters << "All files (*)";
-	
-	
+
+
 	ZDLNameInput diag(this);
 	diag.setWindowTitle("Add Source Port/Engine");
 	diag.setFilter(filters);
@@ -103,8 +103,8 @@ void ZDLSourcePortList::addButton(){
 		ZDLNameListable *zList = new ZDLNameListable(pList, 1001, fileName, name);
 		insert(zList, -1);
 	}
-	
-	
+
+
 }
 
 void ZDLSourcePortList::editButton(QListWidgetItem * item){
@@ -122,6 +122,6 @@ void ZDLSourcePortList::editButton(QListWidgetItem * item){
 			QString name = diag.getName();
 			zitem->setDisplayName(name);
 			zitem->setFile(fileName);
-		}	
+		}
 	}
 }
